View: Add dragging for mouse
This patch adds dragging for mouse, so by pressing middle mouse button
dragging mode is being changed to scroll hand drag, basically it allows
to navigate on the image without pressing scroll bars.

To do that, View class had to be introduced which has been introduced to
override some of QGraphicsView mouse events and to promote QGraphicsView
widget to it.