Config: Open in home directory if app is being run for the 1st time
Previously, if user didn't have config files (which is (most of the
times) equivalent to running the app for the first time), then by
default next file that he wanted to open would be opened in the config
directory. This oneliner changes it, opening it in home directory by
default.