LevelCelView: Fix frame not being displayed properly in Tile/Megatile
This patch fixes the incorrect behaviour where we click on tile/megatile,
selecting a certain frame, and then after pressing
"next/previous/first/last frame button" frame in the frame preview is not the
same one displayed in tile/megatile view.

This is caused by the celFrameIndices QList in D1Min class, which holds
frame indices starting from 1, not 0. This patch adds 1 to all the frame
indices when they are selected from QList in D1Min class, so the offset
is the same one as it is in LevelCelView.