MenuEdit: Do not use autoconnection mechanism for slots used in QMenu (#132)
* MainWindow: Do not use autoconnection mechanism for slots used in QMenu

Currently whenever UI gets loaded and we try to connect a slot with a
signal emitted in QAction, for example in appending frame, inserting it,
etc. we get warnings like these:

qt.core.qmetaobject.connectslotsbyname: QMetaObject::connectSlotsByName: No matching signal for on_actionAdd_Frame_triggered()

This is because MOC tries to automatically connect functions which have
"on_objectname_triggered" signature, and it can't find a matching signal
for it by default. So the solution for that is to simply not use the
autoconnection mechanism by changing names of those slots, removing
"on" prefix from them.

* PaletteWidget: Do not use autoconnection mechanism for slots used in
QMenu

Currently whenever UI gets loaded and we try to connect a slot with a
signal emitted in QAction, for example in appending frame, inserting it,
etc. we get warnings like these:

qt.core.qmetaobject.connectslotsbyname: QMetaObject::connectSlotsByName: No matching signal for on_actionAdd_Frame_triggered()

This is because MOC tries to automatically connect functions which have
"on_objectname_triggered" signature, and it can't find a matching signal
for it by default. So the solution for that is to simply not use the
autoconnection mechanism by changing names of those slots, removing
"on" prefix from them.