Everywhere: Move source files to separate subdirs
Currently all of the source files are located in src/ subdirectory,
although it gets increasingly harder and harder to find anything in this
pile.

This patch moves those files to separate subdirectories that are named
accordingly to the theme that files in them represent, for example -
subdirectory "undostack" will hold undostack and all commands related to
the stack.

There might be code in some places, for example in palettewidget.cpp
that has "undostack" command inside of it, but moving that will come in
a separate refactor PR in the future.