Undostack: Fix segfault when trying to fetch macro on -1 index
Currently, if user adds a command to the undostack, and then undos it,
and then adds another command - this causes a segfault while fetching
macro ID from command on current undoPos. Current undo position pointer
of undostack is being moved to -1 if user undos every possible command
on the stack. Everytime user adds a new command, eraseRedundantCmds()
method is being called, because all commands that are in "undo" state
need to be cleared. So the solution to that is to simply clear every
single macro and command while erasing redundant commands, if the
current undo position is -1 - this basically means that every single
command/macro is in undo state, so we clear all of them.