MW/PaletteWidget: Move palette/translation handling to PaletteWidget
Currently we are splitting the palette/translation handling between `MainWindow` class and `PaletteWidget`, so this patch changes that by moving the whole logic for handling color palettes that was contained in MainWindow to PaletteWidget and unifies it, so all of the handling happens here, instead of MainWindow.

I also found a lot of duplicated logic in MainWindow for translation palletes, uniq translation and normal palettes, I've tried to unify it as much as I could.